update test script class name to start with 'Test'
