+++ SmokeTest/TestCreateUpdateDeleteProgram.py
@@ -13,10 +13,10 @@ from helperRecip.WebdriverUtilities import WebdriverUtilities
 from helperRecip.testcase import *
 
 
-class CreateUpdateDeleteSystem(WebDriverTestCase):
+class TestCreateUpdateDeleteProgram(WebDriverTestCase):
 
-    def testCreateUpdateDeleteSystem(self):
-        self.testname="CreateUpdateDeleteSystem"
+    def testCreateUpdateDeleteProgram(self):
+        self.testname="TestCreateUpdateDeleteProgram"
         self.setup()
         util = WebdriverUtilities()
         util.setDriver(self.driver)
@@ -27,16 +27,15 @@ class CreateUpdateDeleteSystem(WebDriverTestCase):
         myUtil = do.getUtils()
         do.login()
         
-        aEmail = "testrecip@gmail.com" #already exists in the database
-        
-        last_created_object_link = do.createObject("System")
+        last_created_object_link = do.createObject("Program")
         object_name = str(do.util.getTextFromXpathString(last_created_object_link)).strip()
-        do.navigateToObjectAndOpenObjectEditWindow("System", last_created_object_link)
-        do.populateObjectInEditWindow(object_name , grcobject.system_elements, grcobject.system_values, aEmail)
+        do.navigateToObjectAndOpenObjectEditWindow("Program", last_created_object_link)
+        do.populateObjectInEditWindow(object_name , grcobject.program_elements, grcobject.program_values)
         do.openObjectEditWindow()
         do.showHiddenValues()
-        do.verifyObjectValues(grcobject.system_elements, grcobject.system_values)
+        do.verifyObjectValues(grcobject.program_elements, grcobject.program_values)
         do.deleteObject()
 
 if __name__ == "__main__":
+    #import sys;sys.argv = ['', 'Test.testName']
     unittest.main()
