+++ SmokeTest/TestCreateUpdateDeleteSystem.py
@@ -13,10 +13,10 @@ from helperRecip.WebdriverUtilities import WebdriverUtilities
 from helperRecip.testcase import *
 
 
-class CreateUpdateDeleteProgram(WebDriverTestCase):
+class TestCreateUpdateDeleteSystem(WebDriverTestCase):
 
-    def testCreateUpdateDeleteProgram(self):
-        self.testname="CreateUpdateDeleteProgram"
+    def testCreateUpdateDeleteSystem(self):
+        self.testname="TestCreateUpdateDeleteSystem"
         self.setup()
         util = WebdriverUtilities()
         util.setDriver(self.driver)
@@ -27,15 +27,16 @@ class CreateUpdateDeleteProgram(WebDriverTestCase):
         myUtil = do.getUtils()
         do.login()
         
-        last_created_object_link = do.createObject("Program")
+        aEmail = "testrecip@gmail.com" #already exists in the database
+        
+        last_created_object_link = do.createObject("System")
         object_name = str(do.util.getTextFromXpathString(last_created_object_link)).strip()
-        do.navigateToObjectAndOpenObjectEditWindow("Program", last_created_object_link)
-        do.populateObjectInEditWindow(object_name , grcobject.program_elements, grcobject.program_values)
+        do.navigateToObjectAndOpenObjectEditWindow("System", last_created_object_link)
+        do.populateObjectInEditWindow(object_name , grcobject.system_elements, grcobject.system_values, aEmail)
         do.openObjectEditWindow()
         do.showHiddenValues()
-        do.verifyObjectValues(grcobject.program_elements, grcobject.program_values)
+        do.verifyObjectValues(grcobject.system_elements, grcobject.system_values)
         do.deleteObject()
 
 if __name__ == "__main__":
-    #import sys;sys.argv = ['', 'Test.testName']
     unittest.main()
